Use a Twitter list for notifications
This is less awkward than editing a .env file every time you need to
make a change. We query the list members when we start up. A future
improvement would be to watch for list changes, refresh the list
members, and relaunch the stream with new members.

We're also now using the streaming API. This gets around rate-limiting
and gives us tweets faster. Yay!

Due to API limitations, we have to make a REST call to get list members,
so we have to maintain two clients in the code base. 😕